LayoutTests/media/video-seek-after-end-expected.txt

Test that we can seek after reached the end of the video.


EVENT(playing)
EVENT(ended)
EVENT(seeked)
EVENT(playing)
EXPECTED (video.currentTime.toFixed(2) == '3') OK
END OF TEST
